feat/fix(js): monkey patch chroma to support interpolation in jch/jab/hsluv
jch, jab, hsluv are not supported by chroma-js.
if the scale is not smoothed, the code delegates to `chroma.scale`.
however, if the scale is supposed to be over a mode not supported,
the behavior was not as expected.
this applies a monkey patch written by adobe in leonardo to add support for
chroma to interpolate over jch, jab, and hsluv